requirements.txt
streamlit>=1.36
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.2.0
rank_bm25>=0.2.2
numpy>=1.26
pandas>=2.2
tiktoken>=0.7.0
openai>=1.40.0
python-dotenv>=1.0.1

